64位修正 PLVertex结构修改

+++ HelloPanoramaGL/PanoramaGL/PLStructs.h
@@ -28,7 +28,7 @@ typedef struct PLRange PLRange;
 
 struct PLVertex 
 {
-	CGFloat x, y, z;
+	GLfloat x, y, z;
 };
 typedef struct PLVertex PLVertex;
 typedef struct PLVertex PLPosition;
